+++ scout/adapter/mongo/case_events.py
@@ -1,10 +1,10 @@
 import logging
 from collections import Counter
-from typing import Optional
+from typing import Dict, List, Optional
 
 import pymongo
 
-from scout.constants import CASE_STATUSES
+from scout.constants import CASE_STATUSES, CASE_TAGS
 
 LOG = logging.getLogger(__name__)
 
@@ -96,7 +96,6 @@ class CaseEventHandler(object):
             {"$pull": {"assignees": user["_id"]}, "$set": {"status": case["status"]}},
             return_document=pymongo.ReturnDocument.AFTER,
         )
-        LOG.debug("Case updated")
         return updated_case
 
     def update_status(self, institute, case, user, status, link):
@@ -147,7 +146,53 @@ class CaseEventHandler(object):
             {"$set": {"status": status}},
             return_document=pymongo.ReturnDocument.AFTER,
         )
-        LOG.debug("Case updated")
+
+        return updated_case
+
+    def tag_case(self, institute: Dict, case: Dict, user: Dict, tags: List[str], link: str):
+        """Update the status tags of a case.
+
+        This function will create an Event to log that a user have updated the
+        status of a case. Also the status of the case will be updated.
+        Status could be anyone of:
+            ("prioritized", "inactive", "active", "solved", "archived")
+
+        Args:
+            institute (dict): A Institute object
+            case (dict): A Case object
+            user (dict): A User object
+            tag (str): The new tags of the case
+            link (str): The url to be used in the event
+
+        Returns:
+            updated_case
+        """
+
+        for tag in tags:
+            if tag not in CASE_TAGS.keys():
+                LOG.warning("Tag {0} is invalid".format(tag))
+                return None
+
+        LOG.info(
+            "Creating event for updating tags of {0} to {1}".format(case["display_name"], tags)
+        )
+
+        self.create_event(
+            institute=institute,
+            case=case,
+            user=user,
+            link=link,
+            category="case",
+            verb="tag",
+            subject=case["display_name"],
+        )
+
+        LOG.info("Tagging {0} with {1}".format(case["display_name"], tag))
+        updated_case = self.case_collection.find_one_and_update(
+            {"_id": case["_id"]},
+            {"$set": {"tags": tags}},
+            return_document=pymongo.ReturnDocument.AFTER,
+        )
 
         return updated_case
 
@@ -188,7 +233,6 @@ class CaseEventHandler(object):
             {"$set": {"synopsis": content}},
             return_document=pymongo.ReturnDocument.AFTER,
         )
-        LOG.debug("Case updated")
         return updated_case
 
     def archive_case(self, institute, case, user, link):
@@ -222,7 +266,6 @@ class CaseEventHandler(object):
             {"$set": {"status": "archived"}},
             return_document=pymongo.ReturnDocument.AFTER,
         )
-        LOG.debug("Case updated")
         return updated_case
 
     def remove_variants_event(
@@ -278,7 +321,6 @@ class CaseEventHandler(object):
             {"$set": {"research_requested": True}},
             return_document=pymongo.ReturnDocument.AFTER,
         )
-        LOG.debug("Case updated")
         return updated_case
 
     def reset_research(self, institute, case, user, link):
@@ -313,7 +355,6 @@ class CaseEventHandler(object):
             {"$set": {"research_requested": False}},
             return_document=pymongo.ReturnDocument.AFTER,
         )
-        LOG.debug("Case updated")
         return updated_case
 
     def case_dismissed_variants(self, institute, case):
@@ -442,7 +483,6 @@ class CaseEventHandler(object):
             {"$set": {"rerun_monitoring": True}},
             return_document=pymongo.ReturnDocument.AFTER,
         )
-        LOG.debug("Case updated")
         return updated_case
 
     def unmonitor(self, institute, case, user, link):
@@ -473,7 +513,6 @@ class CaseEventHandler(object):
             {"$set": {"rerun_monitoring": False}},
             return_document=pymongo.ReturnDocument.AFTER,
         )
-        LOG.debug("Case updated")
         return updated_case
 
     def share(self, institute, case, collaborator_id, user, link):
@@ -507,7 +546,6 @@ class CaseEventHandler(object):
             {"$push": {"collaborators": collaborator_id}},
             return_document=pymongo.ReturnDocument.AFTER,
         )
-        LOG.debug("Case updated")
         return updated_case
 
     def unshare(self, institute, case, collaborator_id, user, link):
@@ -542,7 +580,6 @@ class CaseEventHandler(object):
             {"$pull": {"collaborators": collaborator_id}},
             return_document=pymongo.ReturnDocument.AFTER,
         )
-        LOG.debug("Case updated")
         return updated_case
 
     def diagnose(
@@ -636,7 +673,6 @@ class CaseEventHandler(object):
             {"$addToSet": {"cohorts": tag}},
             return_document=pymongo.ReturnDocument.AFTER,
         )
-        LOG.debug("Case updated")
         return updated_case
 
     def remove_cohort(self, institute, case, user, link, tag):
@@ -669,7 +705,6 @@ class CaseEventHandler(object):
             {"$pull": {"cohorts": tag}},
             return_document=pymongo.ReturnDocument.AFTER,
         )
-        LOG.debug("Case updated")
         return updated_case
 
     def update_clinical_filter_hpo(
@@ -703,7 +738,6 @@ class CaseEventHandler(object):
             {"$set": {"hpo_clinical_filter": hpo_clinical_filter}},
             return_document=pymongo.ReturnDocument.AFTER,
         )
-        LOG.debug("Case updated")
         return updated_case
 
     def update_case_group_ids(self, institute_obj, case_obj, user_obj, link, group_ids):
@@ -774,6 +808,5 @@ class CaseEventHandler(object):
             {"$set": {"panels": case_obj["panels"]}},
             return_document=pymongo.ReturnDocument.AFTER,
         )
-        LOG.debug("Case updated")
 
         return updated_case